+++ src/net/sf/mpxj/primavera/PrimaveraPMFileReader.java
@@ -23,6 +23,7 @@
 
 package net.sf.mpxj.primavera;
 
+import java.io.IOException;
 import java.io.InputStream;
 import java.util.Collections;
 import java.util.Date;
@@ -36,13 +37,14 @@ import java.util.Set;
 import javax.xml.bind.JAXBContext;
 import javax.xml.bind.JAXBException;
 import javax.xml.bind.Unmarshaller;
+import javax.xml.bind.UnmarshallerHandler;
 import javax.xml.parsers.ParserConfigurationException;
 import javax.xml.parsers.SAXParser;
 import javax.xml.parsers.SAXParserFactory;
-import javax.xml.transform.sax.SAXSource;
 
 import org.xml.sax.InputSource;
 import org.xml.sax.SAXException;
+import org.xml.sax.XMLFilter;
 import org.xml.sax.XMLReader;
 
 import net.sf.mpxj.AssignmentField;
@@ -72,10 +74,8 @@ import net.sf.mpxj.Task;
 import net.sf.mpxj.TaskField;
 import net.sf.mpxj.TimeUnit;
 import net.sf.mpxj.common.BooleanHelper;
-import net.sf.mpxj.common.CharsetHelper;
 import net.sf.mpxj.common.DateHelper;
 import net.sf.mpxj.common.NumberHelper;
-import net.sf.mpxj.common.ReplaceOnceStream;
 import net.sf.mpxj.listener.ProjectListener;
 import net.sf.mpxj.mpp.CustomFieldValueItem;
 import net.sf.mpxj.primavera.schema.APIBusinessObjects;
@@ -121,12 +121,6 @@ public final class PrimaveraPMFileReader extends AbstractProjectReader
    {
       try
       {
-         //
-         // This is a hack to ensure that the incoming file has a namespace
-         // which JAXB will accept.
-         //
-         InputStream namespaceCorrectedStream = new ReplaceOnceStream(stream, NAMESPACE_REGEX, NAMESPACE_REPLACEMENT, NAMESPACE_SCOPE, CharsetHelper.UTF8);
-
          m_projectFile = new ProjectFile();
          m_eventManager = m_projectFile.getEventManager();
 
@@ -150,7 +144,6 @@ public final class PrimaveraPMFileReader extends AbstractProjectReader
          factory.setNamespaceAware(true);
          SAXParser saxParser = factory.newSAXParser();
          XMLReader xmlReader = saxParser.getXMLReader();
-         SAXSource doc = new SAXSource(xmlReader, new InputSource(namespaceCorrectedStream));
 
          if (CONTEXT == null)
          {
@@ -158,8 +151,12 @@ public final class PrimaveraPMFileReader extends AbstractProjectReader
          }
 
          Unmarshaller unmarshaller = CONTEXT.createUnmarshaller();
-
-         APIBusinessObjects apibo = (APIBusinessObjects) unmarshaller.unmarshal(doc);
+         XMLFilter filter = new NamespaceFilter();
+         filter.setParent(xmlReader);
+         UnmarshallerHandler unmarshallerHandler = unmarshaller.getUnmarshallerHandler();
+         filter.setContentHandler(unmarshallerHandler);
+         filter.parse(new InputSource(stream));
+         APIBusinessObjects apibo = (APIBusinessObjects) unmarshallerHandler.getResult();
 
          List<ProjectType> projects = apibo.getProject();
          ProjectType project = null;
@@ -204,6 +201,11 @@ public final class PrimaveraPMFileReader extends AbstractProjectReader
       }
 
       catch (SAXException ex)
+      {
+         throw new MPXJException("Failed to parse file", ex);
+      }
+
+      catch (IOException ex)
       {
          throw new MPXJException("Failed to parse file", ex);
       }
@@ -964,10 +966,6 @@ public final class PrimaveraPMFileReader extends AbstractProjectReader
    private Map<Integer, Integer> m_clashMap = new HashMap<Integer, Integer>();
    private Map<Integer, ProjectCalendar> m_calMap = new HashMap<Integer, ProjectCalendar>();
 
-   private static final int NAMESPACE_SCOPE = 512;
-   private static final String NAMESPACE_REGEX = "xmlns=\\\".*BusinessObjects\\\"";
-   private static final String NAMESPACE_REPLACEMENT = "xmlns=\"http://xmlns.oracle.com/Primavera/P6/V17.7/API/BusinessObjects\"";
-
    private static final Map<String, net.sf.mpxj.ResourceType> RESOURCE_TYPE_MAP = new HashMap<String, net.sf.mpxj.ResourceType>();
    static
    {
@@ -1013,6 +1011,7 @@ public final class PrimaveraPMFileReader extends AbstractProjectReader
    private static final Map<String, Day> DAY_MAP = new HashMap<String, Day>();
    static
    {
+      // Current PMXML schema
       DAY_MAP.put("Monday", Day.MONDAY);
       DAY_MAP.put("Tuesday", Day.TUESDAY);
       DAY_MAP.put("Wednesday", Day.WEDNESDAY);
@@ -1020,6 +1019,15 @@ public final class PrimaveraPMFileReader extends AbstractProjectReader
       DAY_MAP.put("Friday", Day.FRIDAY);
       DAY_MAP.put("Saturday", Day.SATURDAY);
       DAY_MAP.put("Sunday", Day.SUNDAY);
+
+      // Older (6.2?) schema
+      DAY_MAP.put("1", Day.SUNDAY);
+      DAY_MAP.put("2", Day.MONDAY);
+      DAY_MAP.put("3", Day.TUESDAY);
+      DAY_MAP.put("4", Day.WEDNESDAY);
+      DAY_MAP.put("5", Day.THURSDAY);
+      DAY_MAP.put("6", Day.FRIDAY);
+      DAY_MAP.put("7", Day.SATURDAY);
    }
 
    private static final Map<String, Boolean> MILESTONE_MAP = new HashMap<String, Boolean>();